divided the controllers into the contacts folder
