+++ controllers/contacts/addContact.js
@@ -1,0 +1,9 @@
+const contactsOperations = require('../../model/contacts')
+const { sendSuccessResponse } = require('../../helpers')
+
+const addContact = async (req, res) => {
+  const result = await contactsOperations.addContact(req.body)
+  sendSuccessResponse(res, { data: result }, 201)
+}
+
+module.exports = addContact
